switch some methods to provided ones
Created using spr 1.3.6-beta.1
